% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd.R
\name{dwd_multi}
\alias{dwd_multi}
\title{Download and preprocess monthly weather data from DWD}
\usage{
dwd_multi(
  year,
  month,
  variable = c("air_temperature_max", "air_temperature_mean", "air_temperature_min",
    "drought_index", "evapo_p", "evapo_r", "frost_depth", "precipitation", "soil_moist",
    "soil_temperature_5cm", "sunshine_duration"),
  path
)
}
\arguments{
\item{year}{integer. The year for which the weather data will be downloaded.}

\item{month}{integer. The month for which the weather data will be downloaded. One of 1:12.}

\item{variable}{character. The weather parameter of interest.}

\item{path}{character. The folder where results will be stored.}
}
\description{
Download and preprocess monthly weather data from DWD
}
\author{
Jan Linnenbrink, Marvin Ludwig
}
