% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smb.R
\name{get_smb}
\alias{get_smb}
\title{Get Benchmark}
\usage{
get_smb(year, month, path)
}
\arguments{
\item{year}{integer, one of 2020, 2021, 2022}

\item{month}{integer, one of 1 to 12}

\item{path}{string, where to store downloaded data}
}
\description{
Downloads a stack of standardized raster
}
\author{
Marvin Ludwig
}
