% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R, R/landcover.R
\name{landcover_single}
\alias{landcover_single}
\title{WWF Ecoregions}
\usage{
landcover_single(year = 2018, product, path)

landcover_single(year = 2018, product, path)
}
\arguments{
\item{year}{integer. The year for which the land cover will be downloaded. Currently only 2018 is available.}

\item{path}{character. The folder where results will be stored.}

\item{class}{string. One of "GRA", "IMD" or "TCD".}
}
\description{
Downloads WWF ecoregions as categorial predictor.

Downloads a spatially continuous land cover (grassland, build up, forests)
}
\details{
Land cover is based no the Copernicus land monitoring service with 100m spatial resolution.
It was spatially aggregated to a 1km resolution. Hence, land cover classes are represented by % coverage.
GRA is grasslands, IMD is impervious build-up, TCD is tree cover density i.e. forests.
}
\author{
Marvin Ludwig
}
