% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{noise}
\alias{noise}
\title{Random Noise Raster}
\usage{
noise(fun = runif, seed = 1)
}
\arguments{
\item{fun}{a distribution function, see Distribution}

\item{seed}{numeric, RNG seed}
}
\description{
Generates a raster layer based on a distribution function
}
\author{
Marvin Ludwig
}
