% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spampling.R
\name{sample_clustered}
\alias{sample_clustered}
\title{Sample Clustered}
\usage{
sample_clustered(size, ncluster = 5, radius = 25000)
}
\arguments{
\item{size}{numeric,  number of samples to draw (in total)}

\item{ncluster}{numeric, number of cluster centers}

\item{radius}{numeric, buffer size around the cluster centers to draw samples from}
}
\description{
Spatially clustered sample without replacement
}
\details{
Tweaking `radius` decides the severeness of clustering. Larger `radius` means less severe clusters.
    Clusters can overlap, but samples are unique cells.
}
\author{
Marvin Ludwig
}
