% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spampling.R
\name{sample_random}
\alias{sample_random}
\title{Random Sample}
\usage{
sample_random(size)
}
\arguments{
\item{size}{numeric, number of samples to draw}
}
\description{
Random sample without replacement
}
\details{
Can be seen as a probability sample
}
\author{
Marvin Ludwig
}
