% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate}
\usage{
simulate(scenario = "covariate")
}
\arguments{
\item{scenario}{string, one of "covariate", "prior" or "concept" see details}
}
\description{
Simulate prediction challenges
}
\details{
Simulates model transfer scenarios based on covariate shift, prior shift or concept shift (Ma et al. 2024)
}
