% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spampling.R
\name{spampling}
\alias{spampling}
\title{Spampling}
\usage{
spampling(
  method = "random",
  size = 1000,
  cluster = 10,
  radius = 20000,
  seed = 1,
  benchmark = NULL
)
}
\arguments{
\item{method}{string, can be "random" or "clustered"}

\item{size}{integer, the sample size}

\item{cluster}{integer, number of cluster centers}

\item{radius}{integer, sampling distance around cluster centers}

\item{benchmark}{terra::rast, for feature extraction}
}
\description{
Spatial Sampling
}
\details{
TODO
}
\author{
Marvin Ludwig
}
