% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography.R
\name{topography}
\alias{topography}
\title{Download and project topography data}
\usage{
topography(...)
}
\arguments{
\item{...}{see geodata::elevation_30s}
}
\description{
Download and project topography data
}
\author{
Marvin Ludwig
}
